updated collections.py file
Adds more ways to interact with collections, including being able to list works and series bookmarked to them (which was a nightmare because get_work_from_banner has no clue what a series is), being able to see if a collection has parent or sub-collections, getting the number of pages of works in a collection, getting info like Rules, FAQ, and intro for the collection, and how controlled the ability to add works is (i.e. "(Closed, Moderated)").